+++ core/src/main/java/nl/nn/adapterframework/configuration/Configuration.java
@@ -48,9 +48,9 @@ import nl.nn.adapterframework.util.RunStateEnum;
  */
 public class Configuration {
     protected Logger log = LogUtil.getLogger(this);
-    private ClassLoader classLoader = Thread.currentThread().getContextClassLoader();
+    private ClassLoader configurationClassLoader = Thread.currentThread().getContextClassLoader();
 
-	private boolean autoStart = AppConstants.getInstance().getBoolean("configurations.autoStart", true);
+	private boolean autoStart = AppConstants.getInstance(configurationClassLoader).getBoolean("configurations.autoStart", true);
 
     private AdapterService adapterService;
 
@@ -137,7 +137,7 @@ public class Configuration {
 	}
 
 	public ClassLoader getClassLoader() {
-		return classLoader;
+		return configurationClassLoader;
 	}
 
 	public void setAutoStart(boolean autoStart) {
@@ -146,6 +146,10 @@ public class Configuration {
 
 	public boolean isAutoStart() {
 		return autoStart;
+	}
+
+	public boolean isStubbed() {
+		return ConfigurationUtils.isConfigurationStubbed(getClassLoader());
 	}
 
 	/**
@@ -289,7 +293,7 @@ public class Configuration {
 	}
 
 	public String getClassLoaderType() {
-		return classLoader.getClass().getSimpleName();
+		return configurationClassLoader.getClass().getSimpleName();
 	}
 
 	public void setIbisManager(IbisManager ibisManager) {