+++ core/src/test/java/nl/nn/adapterframework/filesystem/MockFileSystem.java
@@ -83,6 +83,20 @@ public class MockFileSystem<M extends MockFile> extends MockFolder implements IW
 	}
 
 	@Override
+	public M toFile(String folderName, String filename) throws FileSystemException {
+		checkOpen();
+		MockFolder destFolder= folderName==null?this:getFolders().get(folderName);
+		if (destFolder==null) {
+			throw new FileSystemException("folder ["+folderName+"] does not exist");
+		}
+		M result = (M)destFolder.getFiles().get(filename);
+		if (result!=null) {
+			return result;
+		}
+		return (M)new MockFile(filename,destFolder);
+	}
+
+	@Override
 	public Iterator<M> listFiles(String folderName) throws FileSystemException {
 		checkOpen();
 		MockFolder folder=folderName==null?this:getFolders().get(folderName);
@@ -173,7 +187,7 @@ public class MockFileSystem<M extends MockFile> extends MockFolder implements IW
 
 	@Override
 	public String getCanonicalName(M f) throws FileSystemException {
-		checkOpenAndExists(null,f);
+		//checkOpenAndExists(null,f); // cannot check this anymore, canonical name is now used in error messages
 		return f.getName();
 	}